- Fix overallocation of memory

+++ lib/include/motioncam/Decoder.hpp
@@ -57,7 +57,7 @@ namespace motioncam {
         const std::vector<Timestamp>& getFrames() const;
         
         // Load a single frame and its metadata.
-        void loadFrame(const Timestamp timestamp, std::vector<uint16_t>& outData, nlohmann::json& outMetadata);
+        void loadFrame(const Timestamp timestamp, std::vector<uint8_t>& outData, nlohmann::json& outMetadata);
         
         // Audio sample rate
         int audioSampleRateHz() const;
